+++ src/test/java/com/github/ddth/queue/qnd/QndMultithreadPgSQL.java
@@ -18,15 +18,17 @@ public class QndMultithreadPgSQL {
     private static ConcurrentMap<Object, Object> SENT = new ConcurrentHashMap<Object, Object>();
     private static ConcurrentMap<Object, Object> RECEIVE = new ConcurrentHashMap<Object, Object>();
     private static AtomicLong TIMESTAMP = new AtomicLong(0);
-    private static long NUM_ITEMS = 8192;
-    private static int NUM_THREADS = 8;
+    private static long NUM_ITEMS = 1024;
+    private static int NUM_THREADS = 4;
 
     public static void main(String[] args) throws Exception {
         BasicDataSource dataSource = new BasicDataSource();
         dataSource.setDriverClassName("org.postgresql.Driver");
-        dataSource.setUrl("jdbc:postgresql://localhost:5432/temp");
+        dataSource.setUrl("jdbc:postgresql://192.168.1.10:5432/test");
         dataSource.setUsername("test");
         dataSource.setPassword("test");
+        dataSource.setMaxTotal(NUM_THREADS);
+        dataSource.setMaxIdle(NUM_THREADS);
 
         final UniversalJdbcQueue queue = new UniversalJdbcQueue();
         // queue.setMaxRetries(10);